+++ src/App.js
@@ -93,6 +93,7 @@ const defaultSettings = {
     surface: [1, 2, 3, 4, 5, 6, 7, 8],
     runway: [0, 30000],
     onlySim: false,
+    onlySimAlternative: false,
     onlyILS: false,
     excludeMilitary: false
   },
@@ -116,25 +117,25 @@ const defaultSettings = {
     jobsPlanesMax: 10,
     express: true,
     expiration: ''
+  },
+  filters: {
+    type: 'Trip-Only',
+    cargo: ['passengers'],
+    fromIcao: null,
+    toIcao: null,
+    minPax: '',
+    minKg: '',
+    maxPax: '',
+    maxKg: '',
+    minDist: '',
+    maxDist: '',
+    minJobPay: '',
+    minLegPay: '',
+    percentPay: '',
+    direction: ''
   }
 };
 
-const defaultFilters = {
-  type: 'Trip-Only',
-  cargo: ['passengers'],
-  fromIcao: null,
-  toIcao: null,
-  minPax: '',
-  minKg: '',
-  maxPax: '',
-  maxKg: '',
-  minDist: '',
-  maxDist: '',
-  minJobPay: '',
-  minLegPay: '',
-  percentPay: '',
-  direction: ''
-};
 
 const icaos = Object.keys(icaodataSrc);
 
@@ -234,8 +235,19 @@ const storage = new Storage();
 
 function App() {
 
+  const [settings, setSettings] = React.useState(() => {
+    const obj = {};
+    // Cannot use _defaultsDeep, because it messes up with array
+    [defaultSettings, storage.get('settings', {})].forEach(item => {
+      _mergeWith(obj, item, (objectValue, sourceValue) => {
+        return _isArray(sourceValue) ? sourceValue : undefined;
+      });
+    });
+    return obj;
+  });
+
   const [filtersBar, setFiltersBar] = React.useState(false)
-  const [filters, setFilters] = React.useState(defaultFilters);
+  const [filters, setFilters] = React.useState(settings.filters);
   const [table, setTable] = React.useState(false);
   const [updatePopup, setUpdatePopup] = React.useState(false);
   const [settingsPopup, setSettingsPopup] = React.useState(false);
@@ -248,16 +260,6 @@ function App() {
   });
   const [flight, setFlight] = React.useState(() => {
     return transformJobs(storage.get('flight', {}));
-  });
-  const [settings, setSettings] = React.useState(() => {
-    const obj = {};
-    // Cannot use _defaultsDeep, because it messes up with array
-    [defaultSettings, storage.get('settings', {})].forEach(item => {
-      _mergeWith(obj, item, (objectValue, sourceValue) => {
-        return _isArray(sourceValue) ? sourceValue : undefined;
-      });
-    });
-    return obj;
   });
   const [search, setSearch] = React.useState(null);
   const [searchDest, setSearchDest] = React.useState(null);
@@ -676,7 +678,7 @@ function App() {
             icaodata={icaodata}
             actions={actions}
             setSettingsPopup={setSettingsPopup}
-            clear={() => setFilters(defaultFilters)}
+            clear={() => setFilters(settings.filters)}
           />
         }
       </AppBar>