+++ src/Popups/Settings.js
@@ -88,6 +88,7 @@ function SettingSwitch({xs, setting, s, setS, label, ...props}) {
               setS(obj);
             }}
             color="primary"
+            {...props}
           />
         }
         label={label}
@@ -214,6 +215,15 @@ function SettingsPopup(props) {
     ['strict', 'Strict: only load jobs from the exact airports where a plane is available'],
     ['around', 'Area: load jobs from an area around airports where a plane is available']
   ];
+  const jobsTypeOptions = [
+    ['Trip-Only', 'Trip Only'],
+    ['VIP', 'VIP'],
+    ['All-In', 'All In']
+  ];
+  const jobsCargoOptions = [
+    ['passengers', 'Passengers'],
+    ['kg', 'Cargo']
+  ];
 
   const handleClose = () => {
     // Cancel change
@@ -280,6 +290,11 @@ function SettingsPopup(props) {
             <FilterAltIcon />&nbsp;<Typography>Filters settings</Typography>
           </AccordionSummary>
           <AccordionDetails>
+            <Typography variant="body1" sx={{ mb: 2 }}>Default filters (when loading FSE Planner):</Typography>
+            <Grid container spacing={3} sx={{ mb: 3 }}>
+              <SettingSelect s={s} setS={setS} label="Job category" setting='filters.type' options={jobsTypeOptions} xs={6} />
+              <SettingSelect s={s} setS={setS} label="Job type" setting='filters.cargo' options={jobsCargoOptions} multiple={true} xs={6} />
+            </Grid>
             <Grid container spacing={3}>
               <Grid item container xs={9} sx={{ alignContent: 'flex-start '}}>
                 <Typography variant="body1" sx={{ mb: 1 }}>From ICAO settings:</Typography>
@@ -327,6 +342,7 @@ function SettingsPopup(props) {
               <SettingSlider3 s={s} setS={setS} label="Airport longest runway (in feet)" setting="airport.runway" xs={12} />
               <SettingSelect s={s} setS={setS} label="Airport runway surface" setting="airport.surface" options={surfaceOptions} multiple={true} xs={12} />
               <SettingSwitch s={s} setS={setS} label="Only display and use simulator compatible airports" setting="airport.onlySim" xs={12} />
+              <SettingSwitch s={s} setS={setS} label="Include non-compatible airports that have at least one alternative in your simulator" setting="airport.onlySimAlternative" xs={12} disabled={!s.airport.onlySim} />
               <SettingSwitch s={s} setS={setS} label="Only display and use airports with an ILS approach (MSFS)" setting="airport.onlyILS" xs={12} />
               <SettingSwitch s={s} setS={setS} label="Exclude military airbases" setting="airport.excludeMilitary" xs={12} />
             </Grid>